[SSTTP2-372] Change time-to-pay-tax-payer microservice to only have the "surname" field as mandatory
